Initial commit
There are 2 folder right now:
itt-local-bot - it has almost all the commands and the datastore
connections. It’s not using cloud functions because it’s what I was
building while waiting for credentials. It has anyway the behaviour I
will merge in the main project if GC Functions performs well.

itt-telegram-js: basically the start command and the serverless.yml
test file to make the gc functions work. It does not save the id on the
cloud so far, but it’s on top of the TODO list.